# public/swagger-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Capstone - API - Documentation",
    "description": "API for user interface development",
    "version": "1.1",
    "contact": {
      "name": "mangox API Support",
      "email": "support@mangox.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/front/layout": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_front_layout",
        "tags": [
          "Front"
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/api/v1/front/sitemap/page": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_front_sitemap_page",
        "tags": [
          "Front"
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/api/v1/front/sitemap/post": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_front_sitemap_post",
        "tags": [
          "Front"
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/api/v1/front/sitemap/post-v2": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "post__api_v1_front_sitemap_post_v2",
        "tags": [
          "Front"
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/api/v1/front/sitemap/blocks": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_front_sitemap_blocks",
        "tags": [
          "Front"
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/api/v1/front/form-builder-content/{slug}": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "post__api_v1_front_form_builder_content__slug_",
        "tags": [
          "Front"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://mangox-api.mangoads.com.vn",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:5555",
      "description": "Development Server"
    }
  ],
...
  "security": [
    {},
    {
      "bearerAuth": []
    }
  ]
}